+++ src/main/java/ua/com/juja/microservices/teams/slackbot/exceptions/ExceptionsHandler.java
@@ -5,17 +5,10 @@ import me.ramswaroop.jbot.core.slack.models.RichMessage;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.client.RestTemplate;
-import ua.com.juja.microservices.teams.slackbot.model.User;
 import ua.com.juja.microservices.teams.slackbot.service.UserService;
-import ua.com.juja.microservices.teams.slackbot.util.Utils;
+import ua.com.juja.microservices.teams.slackbot.util.SlackNameHandler;
 
 import javax.inject.Inject;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * @author Ivan Shapovalov
@@ -43,19 +36,19 @@ public class ExceptionsHandler {
     @ExceptionHandler(Exception.class)
     public void handleAllOtherExceptions(Exception ex) {
         log.warn("Other Exception': {}", ex.getMessage());
-        sendPostResponseAsRichMessage(new RichMessage(ex.getMessage()));
+        sendErrorResponseAsRichMessage(new RichMessage(ex.getMessage()));
     }
 
     @ExceptionHandler(WrongCommandFormatException.class)
     public void handleWrongCommandFormatException(Exception ex) {
         log.warn("WrongCommandFormatException: {}", ex.getMessage());
-        sendPostResponseAsRichMessage(new RichMessage(ex.getMessage()));
+        sendErrorResponseAsRichMessage(new RichMessage(ex.getMessage()));
     }
 
     @ExceptionHandler(UserExchangeException.class)
     public void handleUserExchangeException(UserExchangeException ex) {
         log.warn("UserExchangeException: {}", ex.detailMessage());
-        sendPostResponseAsRichMessage(new RichMessage(ex.getMessage()));
+        sendErrorResponseAsRichMessage(new RichMessage(ex.getMessage()));
     }
 
     @ExceptionHandler(TeamExchangeException.class)
@@ -64,32 +57,12 @@ public class ExceptionsHandler {
         String message = ex.getMessage();
         ApiError apiError = ex.getError();
         if (apiError != null && apiError.getExceptionMessage().contains("#")) {
-            message = replaceUuidsBySlackNamesInMessage(apiError.getExceptionMessage());
+            message = SlackNameHandler.replaceUuidsBySlackNamesInMessage(userService, apiError.getExceptionMessage());
         }
-        sendPostResponseAsRichMessage(new RichMessage(message));
+        sendErrorResponseAsRichMessage(new RichMessage(message));
     }
 
-    private String replaceUuidsBySlackNamesInMessage(String message) {
-        log.debug("Start find and replace uuids by slackNames in message {}", message);
-        String[] messageParts = message.split("#");
-        if (messageParts.length > 1) {
-            Set<String> uuids = new HashSet<>(Arrays.asList(messageParts[1].split(",")));
-            Set<String> slackNames = new HashSet<>();
-            try {
-                List<User> users = userService.findUsersByUuids(new ArrayList<>(uuids));
-                slackNames = users.stream().map(User::getSlack)
-                        .collect(Collectors.toSet());
-            } catch (Exception ex) {
-                log.warn("Nested exception : '{}'", ex.getMessage());
-            }
-            messageParts[1] = Utils.listToStringWithDelimeter(slackNames, ",");
-            message = Utils.arrayToStringWithDelimeter(messageParts, "");
-        }
-        log.debug("Finished find and replace uuids by slackNames in message {}", message);
-        return message;
-    }
-
-    private void sendPostResponseAsRichMessage(RichMessage richMessage) {
+    private void sendErrorResponseAsRichMessage(RichMessage richMessage) {
         try {
             restTemplate.postForObject(responseUrl, richMessage, String.class);
         } catch (Exception ex) {
