+++ src/main/java/ua/com/juja/microservices/teams/slackbot/util/SlackNameHandler.java
@@ -1,17 +1,21 @@
 package ua.com.juja.microservices.teams.slackbot.util;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import ua.com.juja.microservices.teams.slackbot.model.User;
+import ua.com.juja.microservices.teams.slackbot.service.UserService;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 /**
  * @author Ivan Shapovalov
  */
-@Service
 @Slf4j
 public class SlackNameHandler {
 
@@ -31,9 +35,39 @@ public class SlackNameHandler {
         while (matcher.find()) {
             slackNames.add(matcher.group());
         }
-        log.debug("Finished extract slackNames '{}' from text '{}", slackNames.toString(), text);
         log.info("Extracted slackNames '{}' from text '{}", slackNames.toString(), text);
         return slackNames;
+    }
+
+    public static void addAtToSlackNames(List<String> slackNames) {
+        for (int i = 0; i < slackNames.size(); i++) {
+            if (!slackNames.get(i).startsWith("@")) {
+                log.debug("Add '@' to slackName : '{}'", slackNames.get(i));
+                String slackName = slackNames.get(i);
+                slackNames.set(i, "@" + slackName);
+            }
+        }
+    }
+
+    public static String replaceUuidsBySlackNamesInMessage(UserService userService, String message) {
+        log.debug("Start find and replace uuids by slackNames in message {}", message);
+        String messageDelimeter = "#";
+        String[] messageParts = message.split(messageDelimeter);
+        if (messageParts.length > 1) {
+            Set<String> uuids = new HashSet<>(Arrays.asList(messageParts[1].split(",")));
+            Set<String> slackNames = new HashSet<>();
+            try {
+                List<User> users = userService.findUsersByUuids(new ArrayList<>(uuids));
+                slackNames = users.stream().map(User::getSlack)
+                        .collect(Collectors.toSet());
+            } catch (Exception ex) {
+                log.warn("Nested exception : '{}'", ex.getMessage());
+            }
+            messageParts[1] = Utils.listToStringWithDelimeter(slackNames, ",");
+            message = Utils.arrayToStringWithDelimeter(messageParts, "");
+        }
+        log.debug("Finished find and replace uuids by slackNames in message {}", message);
+        return message;
     }
 }
